Re-arrange tests to match source directory structure
Following some discussion in Slack (see
https://botbot.me/freenode/hypothes.is/msg/71389150/) we decided that
there was a consensus in favor of organizing all tests under a top-level
tests/ dir.
