+++ h/static/scripts/tests/base/settings-test.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var settings = require('../base/settings');
+var settings = require('../../base/settings');
 
 function createJSONScriptTag(obj, className) {
   var el = document.createElement('script');
