+++ h/static/scripts/tests/controllers/upgrade-elements-test.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var upgradeElements = require('../controllers/upgrade-elements');
+var upgradeElements = require('../../controllers/upgrade-elements');
 
 describe('upgradeElements', function () {
   it('should upgrade elements matching selectors', function () {
